Capture the SIGINT, SIGTERM, SIGQUIT events sent from the terminal separately to each node.js process
- Sending <CTRL+C> to the main process stdin ('iCubTelemVizServer') is propagated to the 'openmctStaticServer' process.
- Sending any of the three events through the command 'kill' is not propagated from 'iCubTelemVizServer' to 'openmctStaticServer'. Kill has to be sent separately to each process.

+++ iCubTelemVizServer/openMctServerHandler.js
@@ -44,8 +44,12 @@ OpenMctServerHandler.prototype.start = function () {
 }
 
 OpenMctServerHandler.prototype.stop = function () {
-    this.processHandle.kill();
-    return {status: 'PROMISE', message: 'Opem-MCT static server process stopping...'};
+    if (this.processHandle) {
+        this.processHandle.kill();
+        return {status: 'WRPLY', message: 'Process OpenMCT Server stopping...'};
+    } else {
+        return {status: 'WARNING', message: 'No process OpenMCT Server running...'};
+    }
 }
 
 OpenMctServerHandler.prototype.isOn = function () {
